Add files via upload
Medication  remainder system is a game changer in current scenarios ,because  as people age falls they forgot to take their medications on time ,as we can see 45% people of age 65 are  loosing  their memory and forget to take their medicine, and some people cant take medicine on thier own, and some times they got confused which tablets they need to take  to over come this we need a medication remainder and alerting system so that when ever it is time to take medicine then the device will give a alert to the patient until he takes medicine, to make this model available to everyone and easy to use we are implementing a hardware system which will work even without power supply.


WORKING:
In this model the the sensors and everything is connected to the Arduino UNO and power supply is given.
It consists of three push buttons one is to setup the timer and one is to increase the time and another button is to reset. And this time setting process works based on RTC DS3231 module 
And it consists of LED’S and buzzer so whenever it is time to take particular medication the LED will glow at that slot and buzzer will beep as a reminder for the patient to take medicine
And this process will display on the lcd screen of the medicine box ,And the brightness of lcd is controlled  using potentiometer.